+++ generators/app/index.js
@@ -7,13 +7,14 @@ module.exports = generators.Base.extend({
     constructor: function () {
         generators.Base.apply(this, arguments)
 
-        this.config.set('coreVersion', 'generator-speedseed version 0.8.1')
-        global.ss.props.coreVersion = this.config.get('coreVersion')
+        this.config.set('coreVersion', 'generator-speedseed version 0.9.0')
+        this.config.set('component', '')
 
         console.log(this.config.get('coreVersion'))
     },
 
     end() {
+        this.composeWith('speedseed:project')
         this.composeWith('speedseed:template')
         this.composeWith('speedseed:framework')
         this.composeWith('speedseed:css')