+++ generators/seed/core/tasks/webserver.js
@@ -2,19 +2,29 @@ module.exports = ($) => {
     'use strict'
 
     $.gulp.task('webserver', () => {
-        const connect = require('connect-livereload')
+        const browserSync = require('browser-sync').create()
+        $.reload = browserSync.reload
+
+        browserSync.init({
+            open: ($.config.open !== false) ? 'local' : false,
+            port: $.config.port,
+
+            server: {
+                baseDir: $.build.dir
+            },
+
+            ui: {
+              port: $.server.portReload
+            }
+        })
+
         const express = require('express')
-        const open = require('open')
         const request = require('request')
 
         const app = express()
 
-        const PORT = $.config.port
-        const PORT_RELOAD = $.server.portReload
-
         // ROUTES
         app
-        .use(connect({port: PORT_RELOAD}))
         .use(express.static($.build.dir))
         .use('/*', (req, res) => res.sendFile($.path.resolve(__dirname, `../../${$.build.dir}`)))
         .use(`/${$.server.request}`, (req, res) => {
@@ -26,14 +36,5 @@ module.exports = ($) => {
             )
             .pipe(res)
         })
-        .listen(PORT, () => console.log('Listening on port %d', PORT))
-
-        // LIVERELOAD
-        $.tinylr.listen(PORT_RELOAD, () => console.log('Listening on port %d', PORT_RELOAD))
-
-        // LAUNCH
-        if ($.config.open !== false) {
-            open(`http://localhost:${PORT}`)
-        }
     })
 }