+++ generators/seed/template/multi-tic-tac-toe/angularjs/app/components/table/main.js
@@ -1,0 +1,9 @@
+angular
+.module('speedseed')
+.component('ssTable', {
+    template: `{%= include('app/components/table', 'tpl') %}`,
+
+    controller() {
+        this.xo = ''
+    }
+})