+++ generators/seed/template/multi-tic-tac-toe/angularjs/component/main.js
@@ -1,0 +1,8 @@
+angular
+.module('<%= project %>')
+.component('<%= component %>', {
+    template: `{%= include('app/components/<%= component %>', 'tpl') %}`,
+
+    controller() {
+    }
+})