+++ generators/seed/template/multi-tic-tac-toe/lib/angularjs/components/table/main.js
@@ -1,9 +1,0 @@
-angular
-.module('speedseed')
-.component('ssTable', {
-    template: `<%= include('app/components/table', 'tpl') %>`,
-
-    controller() {
-        this.xo = ''
-    }
-})