+++ generators/seed/template/todomvc/lib/jquery/__global.js
@@ -1,3 +1,0 @@
-module.exports = {
-    helloGlobal: 'hello world global'
-}